Silence warnings in math task
